+++ client/js/initialise.js
@@ -1,0 +1,3 @@
+import renderQuiz from "./quiz";
+
+renderQuiz()
